fixed machete-client package, moved models to shared/models common dir
