+++ package.json
@@ -44,7 +44,7 @@
     "file-saver": "^1.3.8",
     "jquery": "^3.6.0",
     "luxon": "^2.0.2",
-    "machete-client": "^1.0.0",
+    "machete-client": "^1.0.2",
     "nanoscroller": "0.8.7",
     "ngx-print": "^1.2.1",
     "npm": "^6.14.13",
