+++ src/app/shared/models/config.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import { ConfigVM } from "machete-client";
 import { Record } from "./record";
-export class Config extends Record<Config> {
+export class Config extends Record<Config> implements ConfigVM {
   key: string;
   value: string;
   description: string;
