+++ src/app/shared/models/worker.ts
@@ -1,6 +1,7 @@
+import { WorkerVM } from "machete-client";
 import { Record } from "./record";
 
-export class Worker extends Record<Worker> {
+export class Worker extends Record<Worker> implements WorkerVM {
   firstname1: string;
   firstname2?: null;
   lastname1: string;
@@ -15,4 +16,6 @@ export class Worker extends Record<Worker> {
   memberStatusID: number;
   skillCodes: string;
   zipCode: string;
+  typeOfWorkID: number;
+  dateOfMembership: string;
 }
